requirements.txt
easyocr==1.7.2
Pillow==11.2.1
schedule==1.2.2
selenium==4.31.0
